Refactor editor pane
