+++ app/components/editor/CodeMirrorSetup.tsx
@@ -1,0 +1,66 @@
+import { RefObject } from "react";
+import { Button } from "@/components/ui/button";
+import { Maximize2, Minimize2, Wand2 } from "lucide-react";
+
+interface CodeMirrorSetupProps {
+  containerRef: RefObject<HTMLDivElement | null>;
+  onToggleFullscreen: () => void;
+  fullScreenEditor: boolean;
+  formatQuery: () => void;
+}
+
+export default function CodeMirrorSetup({
+  containerRef,
+  onToggleFullscreen,
+  fullScreenEditor,
+  formatQuery,
+}: CodeMirrorSetupProps) {
+  const isMac =
+    typeof navigator !== "undefined" && /Mac/i.test(navigator.platform);
+
+  return (
+    <>
+      <div ref={containerRef} className="flex-1" />
+      <div className="absolute top-10 -right-2 z-50 flex flex-col gap-2">
+        <div className="relative group">
+          <Button
+            variant="ghost"
+            size="icon"
+            onClick={onToggleFullscreen}
+            className="text-green-300 hover:bg-transparent hover:text-green-400 transition-all duration-300"
+            aria-label={
+              fullScreenEditor
+                ? "Exit editor fullscreen"
+                : "Enter editor fullscreen"
+            }
+          >
+            {fullScreenEditor ? (
+              <Minimize2 className="w-5 h-5" />
+            ) : (
+              <Maximize2 className="w-5 h-5" />
+            )}
+          </Button>
+          <div className="absolute top-1 right-8 z-30 hidden md:group-hover:block bg-gray-700 text-white text-xs rounded px-3 py-2 shadow-lg whitespace-nowrap">
+            {fullScreenEditor ? "Exit fullscreen" : "Enter fullscreen"}
+            <div className="absolute top-1/2 -right-1 w-2 h-2 bg-gray-700 rotate-45 -translate-y-1/2" />
+          </div>
+        </div>
+        <div className="relative group">
+          <Button
+            variant="ghost"
+            size="icon"
+            onClick={formatQuery}
+            className="text-blue-300 hover:bg-transparent hover:text-blue-400 transition-all duration-300"
+            aria-label="Format SQL"
+          >
+            <Wand2 className="w-5 h-5" />
+          </Button>
+          <div className="absolute top-1 right-8 z-30 hidden md:group-hover:block bg-gray-700 text-white text-xs rounded px-3 py-2 shadow-lg whitespace-nowrap">
+            Format SQL ({isMac ? "⌘+⇧+F" : "Ctrl+Shift+F"})
+            <div className="absolute top-1/2 -right-1 w-2 h-2 bg-gray-700 rotate-45 -translate-y-1/2" />
+          </div>
+        </div>
+      </div>
+    </>
+  );
+}
