+++ app/utils/restrictedKeywords.ts
@@ -1,0 +1,129 @@
+export const restrictedKeywords = [
+  "INSERT",
+  "UPDATE",
+  "DELETE",
+  "DROP",
+  "ALTER",
+  "CREATE",
+  "TRUNCATE",
+  "GRANT",
+  "REVOKE",
+  "MERGE",
+  "REPLACE",
+  "UPSERT",
+  "EXEC",
+  "EXECUTE",
+  "CALL",
+  "COPY",
+  "LOAD DATA",
+  "BULK INSERT",
+  "INSERT IGNORE",
+  "INSERT ON DUPLICATE KEY UPDATE",
+  "RENAME",
+  "ADD",
+  "MODIFY",
+  "CHANGE",
+  "SET SCHEMA",
+  "COMMENT",
+  "COMMIT",
+  "ROLLBACK",
+  "SAVEPOINT",
+  "BEGIN",
+  "START TRANSACTION",
+  "SET ROLE",
+  "SET SESSION",
+  "OWNERSHIP",
+  "SECURITY",
+  "CREATE FUNCTION",
+  "CREATE PROCEDURE",
+  "ALTER FUNCTION",
+  "ALTER PROCEDURE",
+  "DROP FUNCTION",
+  "DROP PROCEDURE",
+  "DO",
+  "CREATE TRIGGER",
+  "ALTER TRIGGER",
+  "DROP TRIGGER",
+  "CREATE EVENT",
+  "ALTER EVENT",
+  "DROP EVENT",
+  "CREATE DATABASE",
+  "ALTER DATABASE",
+  "DROP DATABASE",
+  "VACUUM",
+  "ANALYZE",
+  "REINDEX",
+  "CLUSTER",
+  "PREPARE",
+  "DECLARE",
+  "LOCK",
+  "UNLOCK",
+  "SET",
+  "REFRESH MATERIALIZED VIEW",
+  "IMPORT FOREIGN SCHEMA",
+  "CREATE INDEX",
+  "ALTER INDEX",
+  "DROP INDEX",
+  "CREATE TYPE",
+  "ALTER TYPE",
+  "DROP TYPE",
+  "CREATE DOMAIN",
+  "ALTER DOMAIN",
+  "DROP DOMAIN",
+  "CREATE SEQUENCE",
+  "ALTER SEQUENCE",
+  "DROP SEQUENCE",
+  "CREATE TABLESPACE",
+  "ALTER TABLESPACE",
+  "DROP TABLESPACE",
+  "CREATE POLICY",
+  "ALTER POLICY",
+  "DROP POLICY",
+  "SET CONSTRAINTS",
+  "ALTER SYSTEM",
+  "DISCARD",
+  "RESET",
+  "LISTEN",
+  "NOTIFY",
+  "LOAD XML",
+  "INSERT DELAYED",
+  "CREATE TABLE SELECT",
+  "ALTER VIEW",
+  "DROP VIEW",
+  "FLUSH",
+  "KILL",
+  "PURGE",
+  "RESET MASTER",
+  "RESET SLAVE",
+  "CREATE USER",
+  "ALTER USER",
+  "DROP USER",
+  "SET PASSWORD",
+  "LOCK TABLES",
+  "UNLOCK TABLES",
+  "ATTACH DATABASE",
+  "DETACH DATABASE",
+  "PRAGMA",
+  "CREATE VIRTUAL TABLE",
+  "CREATE OR REPLACE",
+  "FLASHBACK",
+  "CREATE SYNONYM",
+  "ALTER SYNONYM",
+  "DROP SYNONYM",
+  "CREATE MATERIALIZED VIEW",
+  "ALTER MATERIALIZED VIEW",
+  "DROP MATERIALIZED VIEW",
+  "DBCC",
+  "ALTER AUTHORIZATION",
+  "BACKUP",
+  "RESTORE",
+];
+
+export const containsRestrictedKeywords = (query: string): boolean => {
+  const upperQuery = query.toUpperCase();
+  return restrictedKeywords.some((keyword) =>
+    upperQuery.match(
+      new RegExp(`\\b${keyword.replace(/\s+/g, "\\s+")}\\b`, "i")
+    )
+  );
+};
